+++ src/models/ArrowHeadType.ts
@@ -1,0 +1,1 @@
+export type ArrowHeadType = "start" | "end" | "both";