+++ src/preview/components/SlideControls.tsx
@@ -12,12 +12,14 @@ export interface ISlideControlsProps {
   path?: string;
 }
 
-export const SlideControls: React.FunctionComponent<ISlideControlsProps> = ({
+export const SlideControls: React.FunctionComponent<React.PropsWithChildren<ISlideControlsProps>> = ({
   show,
-  path
+  path,
+  children
 }: React.PropsWithChildren<ISlideControlsProps>) => {
   const [previousEnabled, setPreviousEnabled] = React.useState(false);
   const [isPresentationMode, setIsPresentationMode] = React.useState(false);
+  const [showPosition, setShowPosition] = React.useState(false);
 
   const messageListener = (message: MessageEvent<EventData<any>>) => {
     const { command, payload } = message.data;
@@ -117,7 +119,13 @@ export const SlideControls: React.FunctionComponent<ISlideControlsProps> = ({
 
           <SlideControl title="Next" iconName="arrow-right" action={next} isSlideControl />
         </div>
-        <div className="flex items-center justify-end gap-4">
+        <div className="flex items-center justify-end">
+          {
+            showPosition && (
+              children
+            )
+          }
+          <SlideControl title="Toggle mouse position" className='-rotate-90 hover:bg-[var(--vscode-toolbar-hoverBackground)]' iconName="symbol-ruler" action={() => setShowPosition(prev => !prev)} />
           {
             path && (
               <SlideControl title="Open slide source" iconName="preview" action={openSlideSource} />