+++ src/services/DemoRunner.ts
@@ -168,7 +168,7 @@ export class DemoRunner {
     PresenterView.postMessage(WebViewMessages.toWebview.updatePresentationStarted, DemoRunner.isPresentationMode);
     if (DemoRunner.isPresentationMode) {
       DemoPanel.updateMessage("Presentation mode enabled");
-      await DemoRunner.getDemoFile();
+      await DemoRunner.getDemoFile(undefined, true);
       Preview.postMessage(WebViewMessages.toWebview.updateIsInPresentationMode, true);
     } else {
       DemoPanel.updateMessage();
@@ -188,6 +188,7 @@ export class DemoRunner {
     PresenterView.postMessage(WebViewMessages.toWebview.updateRunningDemos, resetContent);
     DemoRunner.togglePresentationMode(false);
     DemoPanel.update();
+    Preview.close();
   }
 
   /**
@@ -196,6 +197,10 @@ export class DemoRunner {
    * @returns {Promise<void>} A promise that resolves when the demo runner has started.
    */
   private static async start(item: ActionTreeItem | { demoFilePath: string; description: string }): Promise<void> {
+    if (Preview.isListening()) {
+      return;
+    }
+
     const executingFile = await DemoRunner.getExecutedDemoFile();
 
     const demoFile = await DemoRunner.getDemoFile(item);
@@ -1187,9 +1192,13 @@ export class DemoRunner {
   /**
    * Retrieves the demo file associated with the given ActionTreeItem.
    * @param item The ActionTreeItem representing the demo file.
+   * @param triggerFirstDemo A boolean indicating whether to trigger the first demo.
    * @returns A Promise that resolves to an object containing the filePath and demo, or undefined if no demo file is found.
    */
-  private static async getDemoFile(item?: ActionTreeItem): Promise<
+  private static async getDemoFile(
+    item?: ActionTreeItem,
+    triggerFirstDemo: boolean = false
+  ): Promise<
     | {
         filePath: string;
         demo: Demos;
@@ -1236,6 +1245,11 @@ export class DemoRunner {
       executingFile.filePath = demoFilePath;
       executingFile.demo = [];
       await DemoRunner.setExecutedDemoFile(executingFile);
+
+      if (triggerFirstDemo) {
+        await commands.executeCommand(COMMAND.start);
+      }
+
       return {
         filePath: demoFilePath,
         demo: demoFiles[demoFilePath],