+++ tsup.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   entry: ["./src/extension.ts"],
   clean: true,
   format: ["cjs"],
-  outDir: "./out",
+  outDir: "./out/extension",
   external: ["vscode"],
   sourcemap: true,
 });