+++ tsup.wc.config.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "tsup";
+
+export default defineConfig({
+  entry: ["./src/preview/webcomponents/index.ts"],
+  minify: true,
+  platform: "browser",
+  format: "esm",
+  outDir: "./out/webcomponents",
+  sourcemap: true,
+});